refactor(controller): split ProductController into ProductApiController and ProductViewController - separate ProductController into ProductApiController (@RestController) for handling API endpoints and ProductViewController (@Controller) for rendering views to improve code modularity and maintainability
