+++ src/main/kotlin/ecommerce/ProductApiController.kt
@@ -1,21 +1,18 @@
 package ecommerce
 
 import org.springframework.http.ResponseEntity
-import org.springframework.stereotype.Controller
-import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.ResponseBody
+import org.springframework.web.bind.annotation.RestController
 import java.net.URI
 
-@Controller
-class ProductController(private val productRepository: ProductRepository) {
+@RestController
+class ProductApiController(private val productRepository: ProductRepository) {
     @PostMapping("/products")
-    @ResponseBody
     fun create(
         @RequestBody product: Product,
     ): ResponseEntity<Void> {
@@ -24,14 +21,12 @@ class ProductController(private val productRepository: ProductRepository) {
     }
 
     @GetMapping("/products")
-    @ResponseBody
     fun read(): ResponseEntity<List<Product>> {
         val products = productRepository.findAllProducts()
         return ResponseEntity.ok().body(products)
     }
 
     @PutMapping("/products/{id}")
-    @ResponseBody
     fun update(
         @RequestBody newProduct: Product,
         @PathVariable id: Long,
@@ -41,18 +36,10 @@ class ProductController(private val productRepository: ProductRepository) {
     }
 
     @DeleteMapping("/products/{id}")
-    @ResponseBody
     fun delete(
         @PathVariable id: Long,
     ): ResponseEntity<Void> {
         productRepository.delete(id)
         return ResponseEntity.noContent().build()
-    }
-
-    @GetMapping("/admin/products")
-    fun table(model: Model): String {
-        model.addAttribute("products", productRepository.findAllProducts())
-        model.addAttribute("product", Product())
-        return "products"
     }
 }
